feat: Created ContactsCard component
